doc_eval/requirements.txt
langchain
pandas
typer
python-dotenv
pyyaml
jinja2
pytest
langchain-google-genai